refactor: change name of package 'screen' to 'gui'
